+++ examples/service/verify/verification/create/verification_create_example.go
@@ -1,0 +1,41 @@
+package main
+
+import (
+	"log"
+	"os"
+
+	"github.com/RJPearson94/twilio-sdk-go"
+	v2 "github.com/RJPearson94/twilio-sdk-go/service/verify/v2"
+	"github.com/RJPearson94/twilio-sdk-go/service/verify/v2/service/verifications"
+	"github.com/RJPearson94/twilio-sdk-go/session/credentials"
+)
+
+var verifySession *v2.Verify
+
+func init() {
+	creds, err := credentials.New(credentials.Account{
+		Sid:       os.Getenv("TWILIO_ACCOUNT_SID"),
+		AuthToken: os.Getenv("TWILIO_AUTH_TOKEN"),
+	})
+	if err != nil {
+		log.Panicf("%s", err.Error())
+	}
+
+	verifySession = twilio.NewWithCredentials(creds).Verify.V2
+}
+
+func main() {
+	resp, err := verifySession.
+		Service("VAXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXX").
+		Verifications.
+		Create(&verifications.CreateVerificationInput{
+			To:      os.Getenv("DESTINATION_PHONE_NUMBER"),
+			Channel: "sms",
+		})
+
+	if err != nil {
+		log.Panicf("%s", err.Error())
+	}
+
+	log.Printf("SID: %s", resp.Sid)
+}