+++ service/verify/v2/service/api_op_client.go
@@ -5,6 +5,9 @@ import (
 	"github.com/RJPearson94/twilio-sdk-go/client"
 	"github.com/RJPearson94/twilio-sdk-go/service/verify/v2/service/rate_limit"
 	"github.com/RJPearson94/twilio-sdk-go/service/verify/v2/service/rate_limits"
+	"github.com/RJPearson94/twilio-sdk-go/service/verify/v2/service/verification"
+	"github.com/RJPearson94/twilio-sdk-go/service/verify/v2/service/verification_check"
+	"github.com/RJPearson94/twilio-sdk-go/service/verify/v2/service/verifications"
 )
 
 // Client for managing a specific service resource
@@ -14,8 +17,11 @@ type Client struct {
 
 	sid string
 
-	RateLimit  func(string) *rate_limit.Client
-	RateLimits *rate_limits.Client
+	RateLimit         func(string) *rate_limit.Client
+	RateLimits        *rate_limits.Client
+	Verification      func(string) *verification.Client
+	VerificationCheck *verification_check.Client
+	Verifications     *verifications.Client
 }
 
 // ClientProperties are the properties required to manage the service resources
@@ -37,6 +43,18 @@ func New(client *client.Client, properties ClientProperties) *Client {
 			})
 		},
 		RateLimits: rate_limits.New(client, rate_limits.ClientProperties{
+			ServiceSid: properties.Sid,
+		}),
+		Verification: func(verificationSid string) *verification.Client {
+			return verification.New(client, verification.ClientProperties{
+				ServiceSid: properties.Sid,
+				Sid:        verificationSid,
+			})
+		},
+		VerificationCheck: verification_check.New(client, verification_check.ClientProperties{
+			ServiceSid: properties.Sid,
+		}),
+		Verifications: verifications.New(client, verifications.ClientProperties{
 			ServiceSid: properties.Sid,
 		}),
 	}