+++ CHANGELOG.md
@@ -18,6 +18,8 @@ FEATURES
 - **New Resource:** Service
 - **New Resource:** Service Rate Limits
 - **New Resource:** Service Rate Limit Bucket
+- **New Resource:** Verification
+- **New Resource:** Verification Check
 
 ## v0.5.0 (2020-10-16)